fix(feishu): share card-text extractor and fetch full card JSON for merge_forward
Extract the interactive-card text walker into card-content.ts so the
fetched-message read path (send.ts) and the merge_forward sub-message
formatter (bot-content.ts) share one extractor instead of drifting on
whitelisted tags, template variables, i18n locale, and post fallback.

Request card_msg_content_type=user_card_content on the merge_forward
im.message.get so forwarded interactive cards return original card JSON
instead of a degraded payload, mirroring getMessageFeishu.
